Rename Graph implementations.
