+++ src/test/java/nl/tudelft/dnainator/parser/HeaderParserTest.java
@@ -3,7 +3,6 @@ package nl.tudelft.dnainator.parser;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.fail;
 import nl.tudelft.dnainator.parser.exceptions.InvalidHeaderFormatException;
-import nl.tudelft.dnainator.parser.impl.HeaderParser;
 
 import org.junit.Test;
